Remove GraalJS and move test case to core
