+++ core/build.gradle
@@ -95,32 +95,6 @@ dependencies {
             "Tables extension for Markdown renderer"
         )
     )
-
-    BuildUtils.addExternalDependency(
-        project,
-        new ExternalDependency(
-            "org.graalvm.polyglot:js-community:${graalVersion}",
-            "GraalJS",
-            "GraalJS",
-            "https://github.com/graalvm/graaljs",
-            "Universal Permissive License",
-            "https://github.com/graalvm/graaljs/blob/master/LICENSE",
-            "Server-side JavaScript evaluation"
-        )
-    )
-
-    BuildUtils.addExternalDependency(
-        project,
-        new ExternalDependency(
-            "org.graalvm.js:js-scriptengine:${graalVersion}",
-            "GraalJS",
-            "GraalJS",
-            "https://github.com/graalvm/graaljs",
-            "Universal Permissive License",
-            "https://github.com/graalvm/graaljs/blob/master/LICENSE",
-            "Server-side JavaScript evaluation"
-        )
-    )
 }
 
 BuildUtils.addLabKeyDependency(project: project, config: "implementation", depProjectPath: BuildUtils.getBootstrapProjectPath(gradle))
